+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to the "vscode-debug-adapter-apache-camel" extension will be
 - Provide error message and remind requirements in case of failure when trying to attach the debugger
 - Provide configuration snippets for Camel debugger launch configuration
 - Provide configuration snippets to launch Camel application ready to accept a Camel debugger connection using jbang or Maven
+- Upgrade Debug Adapter for Apache Camel to 0.2.0
 
 ## 0.1.1